requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
factorio-rcon-py==2.1.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
py-cord==2.6.1
sniffio==1.3.1
yarl==1.17.0
